% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_col.R
\name{convert_col}
\alias{convert_col}
\title{Generate stable color labels}
\usage{
convert_col(
  data,
  breaks,
  pal_fun,
  include.lowest = TRUE,
  right = TRUE,
  na.col = NA
)
}
\arguments{
\item{data}{Vector of numeric values to be convert into colors}

\item{breaks}{Vector of breaks to use for cutting}

\item{pal_fun}{Vector of colors for the palette}

\item{include.lowest}{Include.lowest for the cut function}

\item{right}{Right option for the cut function}

\item{na.col}{Colour to use in case of NAs}
}
\value{
Returns a list including the colors generated (fill_by), the tags for the intervals (tags) and the colors corresponding for each interval (colors)
}
\description{
This function generates stable color labels to ease comparison when representing several figures that
we would want to have the same palette.
}
