% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.full.graph.R
\name{inla.full.graph}
\alias{inla.full.graph}
\title{Make a fully connected graph for INLA}
\usage{
inla.full.graph(
  sp_obj,
  graph,
  thresh = 0.1,
  snap = 5e-07,
  distance_by = c("perimeter", "centroid")
)
}
\arguments{
\item{sp_obj}{Spatial object in sp form}

\item{graph}{Graph object from nb2INLA function}

\item{snap}{Snap argument for the function poly2nb}

\item{distance_by}{Choices are perimeter and centroid}
}
\value{
List including three different objects: the adjacency object, the edited graph and a dataframe with the changes applied
}
\description{
This function converts a graph with islands into a fully connected graph
by searching for the nearest area between the different nodes and connecting
them, with the possibility of setting a threshold for connecting several areas
if they are within the limits.
}
