% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.MARB.R
\name{inla.MARB}
\alias{inla.MARB}
\title{Calculate MARB between fitted values in INLA and simulated values}
\usage{
inla.MARB(fit_values, sim_values, n.sim = 1)
}
\arguments{
\item{fit_values}{Mean fitted values from INLA models}

\item{sim_values}{Vector of simulated values}

\item{n.sim}{Number of simulated datasets}
}
\value{
Value estimated for MARB
}
\description{
This function estimates Mean Absolute Relative Bias between fitted values in INLA and simulated values
}
