% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.MedARB.R
\name{inla.MedARB}
\alias{inla.MedARB}
\title{Calculate MedARB between fitted values in INLA and simulated values}
\usage{
inla.MedARB(fit_values, sim_values, n.sim = 1)
}
\arguments{
\item{fit_values}{Mean fitted values from INLA models}

\item{sim_values}{Vector of simulated values}

\item{n.sim}{Number of simulated datasets}
}
\value{
Value estimated for MedARB
}
\description{
This function estimates Median Absolute Relative Bias between fitted values in INLA and simulated values
}
