% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.MedRRMSE.R
\name{inla.MedRRMSE}
\alias{inla.MedRRMSE}
\title{Calculate MedRRMSE between fitted values in INLA and simulated values}
\usage{
inla.MedRRMSE(fit_values, sim_values, n.sim = 1)
}
\arguments{
\item{fit_values}{Mean fitted values from INLA models}

\item{sim_values}{Vector of simulated values}

\item{n.sim}{Number of simulated datasets}
}
\value{
Value estimated for MedRRMSE
}
\description{
This function estimates Median Relative Root Mean Prediction Error between fitted values in INLA and simulated values
}
