% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.MRRMSE.R
\name{inla.MRRMSE}
\alias{inla.MRRMSE}
\title{Calculate MRRMSE between fitted values in INLA and simulated values}
\usage{
inla.MRRMSE(fit_values, sim_values, n.sim = 1)
}
\arguments{
\item{fit_values}{Mean fitted values from INLA models}

\item{sim_values}{Vector of simulated values}

\item{n.sim}{Number of simulated datasets}
}
\value{
Value estimated for MRRMSE
}
\description{
This function estimates Mean Relative Root Mean Prediction Error between fitted values in INLA and simulated values
}
