% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.null.sp.R
\name{inla.null.sp}
\alias{inla.null.sp}
\title{Extract null effects from SR-ANOVA models}
\usage{
inla.null.sp(mod, thres = 0.125)
}
\arguments{
\item{mod}{Mod file coming from a SR-ANOVA object}

\item{thres}{Threshold of sd use for the spatial effect to consider it null}
}
\value{
SR-ANOVA object with the summary table updated to include null spatial effects
}
\description{
This function extracts the total number of null spatial effects inside a
SR-ANOVA object, for each one of the different specifications included.
}
