% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.rerun.SpANOVA.R
\name{inla.rerun.SpANOVA}
\alias{inla.rerun.SpANOVA}
\title{Run the desire specification from a SpANOVA object}
\usage{
inla.rerun.SpANOVA(
  SpANOVA_obj,
  n_mod,
  gr,
  scale.mod = TRUE,
  sp.prior = "sdunif",
  pc.prec.val = c(1, 0.01),
  sp.copy.fixed = TRUE,
  verbose.INLA = FALSE
)
}
\arguments{
\item{SpANOVA_obj}{Object generated after running the inla.SpANOVA.2x2 function}

\item{n_mod}{Number of the model that you want to run, using the value reflected in the SpANOVA summary table}

\item{gr}{Graph for the underlying spatial structure}

\item{scale.mod}{Scale copies of random spatial effects or not, default is TRUE}

\item{sp.prior}{Select prior for the random spatial effect, options are sdunif and pc.prec}

\item{pc.prec.val}{Define values por the pc prior in case it was chosen. Default values are c(1, 0.01)}

\item{sp.copy.fixed}{Fix copied values for the random spatial effects, default is TRUE}

\item{verbose.INLA}{Change verbose value for INLA in case you want to follow the approximation}
}
\value{
Object generated by INLA, including groups and sp_effects data from the SpANOVA specification
}
\description{
This function runs the desire specification from a SpANOVA object, using the data prepare from the original specification
}
