% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SpANOVA.R
\name{plot.SpANOVA}
\alias{plot.SpANOVA}
\title{Represent different values from SpANOVA modelling}
\usage{
\method{plot}{SpANOVA}(
  obj,
  obj_type = c("SpANOVA", "INLA"),
  fill_by = c("Spatial", "Heterogeneity", "RR"),
  n_mod,
  sp_obj,
  breaks = NA,
  fil_scale = c("#133BF2", "#7189F7", "#FFFFFF", "#FF867A", "#FF2F1B"),
  col_frontiers = "black",
  scale_name = "Values",
  sp_null = 0.125,
  legend.position = "right",
  ncol_fig = 2
)
}
\arguments{
\item{obj}{Object to generate the plot from}

\item{obj_type}{Type of object provided, options are SpANOVA or INLA after running inla.rerun.SpANOVA}

\item{fill_by}{Values to represent, choosing between Spatial, Heterogeneity and RR}

\item{n_mod}{Number of the model specification that the user wants to represent}

\item{sp_obj}{Spatial object to be plotted}

\item{breaks}{Breaks for the color palette in case the user wants to modify the default ones}

\item{fil_scale}{Vector of colors in case the user wants to modify the default ones}

\item{col_frontiers}{Colour for the lines that define each polygon, default is black}

\item{scale_name}{Name for the filling scale of the plot}

\item{sp_null}{Threshold desired to considered a spatial effect represented null, default is 0.125}

\item{legend.position}{Position for the legend of the plot}

\item{ncol_fig}{Number of columns desired for the grid of figures}
}
\value{
Representation of the values desired using ggplot2 as baseline
}
\description{
This function represents the desired value from a SpANOVA object or an INLA object after running inla.rerun.SpANOVA
}
