% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_sp_ef}
\alias{sim_sp_ef}
\title{Mini OTU abundance dataset}
\format{
A data frame
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.
}
\usage{
sim_sp_ef
}
\description{
A dataset containing relatives abundances of OTUs for human stool samples
with a binary outcome, \code{dx}.
This is a subset of \code{otu_small}.
}
\keyword{datasets}
