% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_object_sim}
\alias{sp_object_sim}
\title{Simulated Datasets}
\format{
Spatial object containing all the simulated datasets for each scenario considered
}
\usage{
sp_object_sim
}
\description{
Spatial object containing all the simulated datasets for each scenario considered
}
\keyword{datasets}
