% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_data_us_23}
\alias{temp_data_us_23}
\title{Mini OTU abundance dataset - preprocessed}
\format{
An object of class \code{data.frame} with 3107 rows and 7 columns.
}
\usage{
temp_data_us_23
}
\description{
This is the result of running \code{preprocess_data("otu_mini_bin")}
}
\keyword{datasets}
