% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_county_23}
\alias{us_county_23}
\title{US County 2023}
\format{
Spatial Object
Spatial object with the US Counties, year 2023
}
\usage{
us_county_23
}
\description{
Spatial object coming from a shp file with the US counties
}
\keyword{datasets}
