% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.SpANOVA.2x2.R
\name{inla.SpANOVA.2x2}
\alias{inla.SpANOVA.2x2}
\title{Apply SpANOVA modelization using a wrapper function in INLA}
\usage{
inla.SpANOVA.2x2(
  obs,
  exp,
  gr,
  fac.names = NULL,
  lev.fac1 = NULL,
  lev.fac2 = NULL,
  scale.mod = TRUE,
  sp.prior = "sdunif",
  pc.prec.val = c(1, 0.01),
  sp.copy.fixed = TRUE,
  save.res = TRUE,
  save.random = TRUE,
  save.hyper = TRUE
)
}
\arguments{
\item{obs}{Vector of observed values}

\item{exp}{Vector of expected values}

\item{gr}{Graph for the underlying spatial structure}

\item{fac.names}{Names of the factors included}

\item{lev.fac1}{Levels of the first factor included}

\item{lev.fac2}{Levels of the second factor included}

\item{scale.mod}{Scale copies of random spatial effects or not, default is TRUE}

\item{sp.prior}{Select prior for the random spatial effect, options are sdunif and pc.prec}

\item{pc.prec.val}{Define values por the pc prior in case it was chosen. Default values are c(1, 0.01)}

\item{sp.copy.fixed}{Fix copied values for the random spatial effects, default is TRUE}

\item{save.res}{Save fitted values or not from the different models, default is TRUE}

\item{save.random}{Save values adjusted from the random spatial effects or not from the different models, default is TRUE}

\item{save.hyper}{Save hyperparameter values from each individual model, default is TRUE}
}
\value{
List with all the models analyzed and a summary table with the most common performance metrics.
}
\description{
This function is a wrapper for multidimensional spatial factor models
in INLA, using a sequential shared spatial effects with nested structure
as discussed in AÑADIR REFERENCIA. \if{html}{\out{<br/>}}
\if{html}{\out{<br/>}}
\if{html}{\out{<b>}}DISCLAIMER:\if{html}{\out{</b>}} Observed and expected values have to be given in an specific order.
Consider n the number of areas, the first n values (1:n) of the obs (exp) should be the ones belonging to the
FIRST level (the first position of the \if{html}{\out{<b>}}lev.fac1\if{html}{\out{</b>}} vector argument) of the
FIRST factor (the first position of the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}} vector argument) and
to the FIRST level (the first position of the \if{html}{\out{<b>}}lev.fac2\if{html}{\out{</b>}} vector argument)
of the SECOND factor (the second position of the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}} vector argument).
The n following values ((n+1):2n) of the obs (exp) should be the ones
belonging to the FIRST level (the first position of the \if{html}{\out{<b>}}lev.fac1\if{html}{\out{</b>}} vector
argument) of the FIRST factor (the first position of the fac.names\if{html}{\out{</b>}}
vector argument) and to the SECOND level (the second position of the
\if{html}{\out{<b>}}lev.fac2\if{html}{\out{</b>}} vector argument) of the SECOND factor (the second position of
the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}} vector argument).\if{html}{\out{<br/>}}
\if{html}{\out{<br/>}}
The n following values ((2n+1):3n) of the obs (exp) should be the ones
belonging to the SECOND level (the second position of the \if{html}{\out{<b>}}lev.fac1\if{html}{\out{</b>}} vector
argument) of the FIRST factor (the first position of the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}}
vector argument) and to the FIRST level (the first position of the
\if{html}{\out{<b>}}lev.fac2\if{html}{\out{</b>}} vector argument) of the SECOND factor (the second position of
the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}} vector argument).\if{html}{\out{<br/>}}
\if{html}{\out{<br/>}}
The n following values ((3n+1):4n) of the obs (exp) should be the ones
belonging to the SECOND level (the second position of the \if{html}{\out{<b>}}lev.fac1\if{html}{\out{</b>}} vector
argument) of the FIRST factor (the first position of the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}}
vector argument) and to the SECOND level (the second position of the
\if{html}{\out{<b>}}lev.fac2\if{html}{\out{</b>}} vector argument) of the SECOND factor (the second position of
the \if{html}{\out{<b>}}fac.names\if{html}{\out{</b>}} vector argument).\if{html}{\out{<br/>}}
\if{html}{\out{<br/>}}
The first n values are O1, the second n values are O2, the third n values
are O3 and the last n values are O4
}
