% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_data_us_23}
\alias{pop_data_us_23}
\title{Population per US County - Year 2023}
\format{
Dataframe
Dataframe containing the population per US county on 2023
}
\usage{
pop_data_us_23
}
\description{
Population per US County - Year 2023
}
\keyword{datasets}
