% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_sp_ef}
\alias{sim_sp_ef}
\title{Simulated spatial effects}
\format{
Matrix
Each column contains one of the simulated spatial effect
}
\usage{
sim_sp_ef
}
\description{
A dataset containing simulated spatial effects.
}
\keyword{datasets}
