% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_data_us_23}
\alias{temp_data_us_23}
\title{Mean Temperature 2023}
\format{
Dataframe containing the mean temperature per US county on 2023
}
\usage{
temp_data_us_23
}
\description{
Mean Temperature per US County - Year 2023
}
\keyword{datasets}
