% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_county_23}
\alias{us_county_23}
\title{Mini OTU abundance dataset with 3 categorical variables}
\format{
A data frame
The \code{dx} column is the colorectal cancer diagnosis: adenoma, carcinoma, normal.
All other columns are OTU relative abundances.
}
\usage{
us_county_23
}
\description{
A dataset containing relatives abundances of OTUs for human stool samples
}
\keyword{datasets}
